src/requirements.txt
dj-database-url==2.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
Faker==20.1.0
gunicorn==21.2.0
parameterized==0.9.0
psycopg2-binary==2.9.9
pymongo==4.6.1
python-decouple==3.8
whitenoise==6.6.0
